Update decide_gi.py
rewrite the decide_gi to be a .test() function which returns bool if the test has passed True or False. The function itself writes to commandline and CSV.
A link to the CSV is presented to the user at the beginning of the test (PyCharm can open this link). Because only faults are logged, the progress is visible by checking the contents of the CSV.